requirements.txt
numpy==1.21.5
Pillow==10.0.0
PyQt6==6.5.2
PyQt6_sip==13.5.2
torch==1.9.1
torchvision==0.10.1
